% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf.R
\docType{data}
\name{geom_sf_label}
\alias{geom_sf_label}
\alias{StatSfCoordinates}
\alias{geom_sf_label_repel}
\title{Variants Of 'geom_sf()'}
\format{An object of class \code{StatSfCoordinates} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatSfCoordinates

geom_sf_label(mapping = NULL, data = NULL,
  fun.geometry = sf::st_point_on_surface, ...)

geom_sf_label_repel(mapping = NULL, data = NULL,
  fun.geometry = sf::st_point_on_surface, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Variants Of 'geom_sf()'
}
\keyword{datasets}
