% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-utils.R
\name{sf-utils}
\alias{sf-utils}
\alias{coordinates_to_column}
\title{Utilities for 'sf'}
\usage{
coordinates_to_column(x, col_names = NULL, fun = sf::st_centroid)
}
\arguments{
\item{x}{sf object.}

\item{col_names}{Column names.}

\item{fun}{Function to calcuate POINT from the geometry.}
}
\description{
Utilities for 'sf'
}
\examples{
points_sfc <- sf::st_sfc(
  sf::st_point(c(1, 1)),
  sf::st_point(c(2, 2))
)

points_sf <- sf::st_sf(
  value = c(10, 20),
  geometry = points_sfc
)

# For POINT, add the coordinates of the points.
coordinates_to_column(points_sf)
coordinates_to_column(points_sf, col_names = c("long", "lat"))

# For other features, add the coordinates of the centroids of them.
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
nc_3857 <- sf::st_transform(nc, 3857)
coordinates_to_column(nc_3857)

# You can specify other functions that return one POINT per geometry.
coordinates_to_column(nc_3857, fun = sf::st_point_on_surface)
}
