% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yutanutils-package.R
\docType{package}
\name{yutanutils-package}
\alias{yutanutils}
\alias{yutanutils-package}
\title{Yutani's Utilities}
\description{
Utilities for me, but maybe also for you.
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com}

}
