% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-utils.R
\name{sf-utils}
\alias{sf-utils}
\alias{st_add_point_coordinates}
\title{Utilities for 'sf'}
\usage{
st_add_point_coordinates(x, col_names = NULL)
}
\arguments{
\item{x}{sf object.}

\item{col_names}{Column names.}
}
\description{
Utilities for 'sf'
}
\examples{
points_sfc <- sf::st_sfc(
  sf::st_point(c(1, 1)),
  sf::st_point(c(2, 2))
)

points_sf <- sf::st_sf(
  value = c(10, 20),
  geometry = points_sfc
)

st_add_point_coordinates(points_sf)
st_add_point_coordinates(points_sf, col_names = c("long", "lat"))
}
